+++ tests/testthat/test_readMSData2.R
@@ -23,18 +23,6 @@ mzf <- .getMzMLFiles(TRUE)[1:2]
 ## Load the data with readMSData mode = onDisk
 odmse <- readMSData(files = mzf, centroided. = TRUE, mode = "onDisk")
 
-test_that("Constructor performance and test for MS1 only", {
-    featDat <- fData(odmse)
-    featDat <- featDat[featDat$fileIdx == 1, ]
-    ## system.time(
-    ##     spR <- MSnbase:::.applyFun2SpectraOfFileSlow(featDat, filenames=fileNames(odmse))
-    ## ) ## 19.5 sec.
-    system.time(
-        spM <- MSnbase:::.applyFun2SpectraOfFileMulti(featDat, filenames=fileNames(odmse))
-    ) ## 3.2 sec.
-    ## expect_equal(spR, spM)
-})
-
 test_that("Constructor performance and test for MSn", {
     ## Get the test data file.
     mzf <- MsDataHub::TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01.20141210.mzML.gz()
@@ -49,12 +37,4 @@ test_that("Constructor performance and test for MSn", {
         spM <- MSnbase:::.applyFun2SpectraOfFileMulti(featDat, filenames=fileNames(odmsn))
     ) ## 1.6 sec.
     ## expect_equal(spR, spM)
-})
-
-test_that("readMSData inMemory and onDisk reading CDF", {
-    library(msdata)
-    f <- system.file("cdf/ko15.CDF",  package = "msdata")
-    odmse <- readMSData(f, mode = "onDisk")
-    mse <- readMSData(f, msLevel. = 1, mode = "inMemory")
-    all.equal(spectra(odmse), spectra(mse))
 })