+++ NEWS.md
@@ -1,7 +1,7 @@
 # MSnbase 2.39
 
 ## MSnbase 2.39.4
--
+- Move more tests to longtests.
 
 ## MSnbase 2.39.3
 - Use long tests.